fix: remove storybook related const from @twreporter/core packages
