+++ packages/core/src/constants/size.js
@@ -6,9 +6,3 @@ export const SIZE = {
 }
 
 export const SIZE_PROP_TYPES = PropTypes.oneOf([SIZE.S, SIZE.L])
-
-export const SIZE_STORYBOOK_ARG_TYPE = {
-  defaultValue: SIZE.S,
-  options: [SIZE.S, SIZE.L],
-  control: { type: 'radio' },
-}
