+++ packages/core/src/constants/theme.js
@@ -14,12 +14,6 @@ export const THEME_PROP_TYPES = PropTypes.oneOf([
   THEME.index,
 ])
 
-export const THEME_STORYBOOK_ARG_TYPE = {
-  defaultValue: THEME.normal,
-  options: [THEME.normal, THEME.photography, THEME.transparent, THEME.index],
-  control: { type: 'radio' },
-}
-
 // the value of ARTICLE_THEME is correspond to posts.style
 export const ARTICLE_THEME = Object.freeze({
   v2: {
