+++ packages/react-components/src/icon/stories/article.stories.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Article } from '../index'
-import { BRANCH_STORYBOOK_ARG_TYPE } from '@twreporter/core/lib/constants/release-branch'
+import { BRANCH_STORYBOOK_ARG_TYPE } from '../../storybook/constants'
 
 export default {
   title: 'Icon/Article',
