+++ packages/react-components/src/icon/stories/text.stories.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Text } from '../index'
-import { BRANCH_STORYBOOK_ARG_TYPE } from '@twreporter/core/lib/constants/release-branch'
+import { BRANCH_STORYBOOK_ARG_TYPE } from '../../storybook/constants'
 
 export default {
   title: 'Icon/Text',
