+++ packages/react-components/src/logo/stories/logoFooter.stories.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import LogoFooter from '../components/logo-footer'
-import { BRANCH_STORYBOOK_ARG_TYPE } from '@twreporter/core/lib/constants/release-branch'
+import { BRANCH_STORYBOOK_ARG_TYPE } from '../../storybook/constants'
 
 export default {
   title: 'Logo/Footer',
