+++ packages/react-components/src/logo/stories/logoSymbol.stories.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import LogoSymbol from '../components/logo-symbol'
-import { BRANCH_STORYBOOK_ARG_TYPE } from '@twreporter/core/lib/constants/release-branch'
+import { BRANCH_STORYBOOK_ARG_TYPE } from '../../storybook/constants'
 
 export default {
   title: 'Logo/Symbol',
