+++ packages/universal-header/src/components/hamburger-footer.stories.js
@@ -3,11 +3,11 @@ import React from 'react'
 import styled from 'styled-components'
 import Footer from './hamburger-footer'
 import HeaderContext from '../contexts/header-context'
+import { THEME } from '@twreporter/core/lib/constants/theme'
 import {
-  THEME,
   THEME_STORYBOOK_ARG_TYPE,
-} from '@twreporter/core/lib/constants/theme'
-import { BRANCH_STORYBOOK_ARG_TYPE } from '@twreporter/core/lib/constants/release-branch'
+  BRANCH_STORYBOOK_ARG_TYPE,
+} from '../storybook/constants'
 
 export default {
   title: 'Hamburger/Footer',
