+++ packages/universal-header/src/components/slogan.stories.js
@@ -2,7 +2,7 @@
 import React from 'react'
 import Slogan from './slogan'
 import HeaderContext from '../contexts/header-context'
-import { THEME_STORYBOOK_ARG_TYPE } from '@twreporter/core/lib/constants/theme'
+import { THEME_STORYBOOK_ARG_TYPE } from '../storybook/constants'
 
 export default {
   title: 'Slogan',
