+++ packages/universal-header/src/components/tab-bar.stories.js
@@ -3,8 +3,10 @@ import React from 'react'
 import styled from 'styled-components'
 import TabBar from './tab-bar'
 import HeaderContext from '../contexts/header-context'
-import { THEME_STORYBOOK_ARG_TYPE } from '@twreporter/core/lib/constants/theme'
-import { BRANCH_STORYBOOK_ARG_TYPE } from '@twreporter/core/lib/constants/release-branch'
+import {
+  THEME_STORYBOOK_ARG_TYPE,
+  BRANCH_STORYBOOK_ARG_TYPE,
+} from '../storybook/constants'
 
 export default {
   title: 'Tab Bar',
