+++ packages/universal-header/src/storybook/utils/get-enum-arg.js
@@ -1,0 +1,7 @@
+export const getRadioArg = (enumObject, defaultValue) => {
+  return {
+    defaultValue,
+    options: Object.values(enumObject),
+    control: { type: 'radio' },
+  }
+}
